Shrink unsuported renderer overlay so test can access valid image data
quickinspectortext was failing on windows because the overlay that
states "DirectX is not supported yet, please use OpenGL or Software
backend" dims color values, making color compare fail. A solution
for this is to make this overlay a little smaller, so we can get a
glimpse of the original colors from the corners we pick colors from
in this test.